+++ s/base/helpers/mixin.ts
@@ -7,6 +7,8 @@ import {Lean} from "../../reactor/types.js"
 import {Flat} from "../../flatstate/flat.js"
 import {BaseElementClass} from "../element.js"
 import {SignalTower} from "../../signals/tower.js"
+import { Context } from "../../nexus/context.js"
+import { Pipe } from "../../tools/pipe.js"
 
 export namespace mixin {
 
@@ -108,6 +110,15 @@ export namespace mixin {
 					}
 				}
 			}
+		}
+	}
+
+	export function setup(...styles: CSSResultGroup[]) {
+		return function<C extends BaseElementClass>(Base: C): C {
+			return Pipe.with(Base)
+				.to(css(...styles))
+				.to(reactor())
+				.done() as C
 		}
 	}
 }